refactor directory heirarchy
