+++ android/src/main/java/com/msopentech/thali/toronionproxy/OnionProxyContext.kt
@@ -10,9 +10,15 @@ MERCHANTABLITY OR NON-INFRINGEMENT.
 
 See the Apache 2 License for the specific language governing permissions and limitations under the License.
 */
-package com.msopentech.thali.universal.toronionproxy
+package com.msopentech.thali.toronionproxy
 
 import android.os.Process
+import com.msopentech.thali.toronionproxy.settings.DefaultSettings
+import com.msopentech.thali.toronionproxy.settings.TorSettings
+import com.msopentech.thali.toronionproxy.settings.TorSettingsBuilder
+import com.msopentech.thali.toronionproxy.util.FileUtilities
+import com.msopentech.thali.toronionproxy.util.TorInstaller
+import com.msopentech.thali.toronionproxy.util.WriteObserver
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import java.io.BufferedWriter
