+++ android/src/main/java/com/msopentech/thali/toronionproxy/OnionProxyManager.kt
@@ -25,13 +25,16 @@ MERCHANTABLITY OR NON-INFRINGEMENT.
 
 See the Apache 2 License for the specific language governing permissions and limitations under the License.
 */
-package com.msopentech.thali.universal.toronionproxy
+package com.msopentech.thali.toronionproxy
 
 import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.Intent
 import android.content.IntentFilter
 import android.net.ConnectivityManager
+import com.msopentech.thali.toronionproxy.broadcaster.DefaultEventBroadcaster
+import com.msopentech.thali.toronionproxy.broadcaster.EventBroadcaster
+import com.msopentech.thali.toronionproxy.util.FileUtilities
 import net.freehaven.tor.control.EventHandler
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
