+++ android/src/main/java/com/msopentech/thali/toronionproxy/OnionProxyManagerEventHandler.kt
@@ -25,7 +25,7 @@ MERCHANTABLITY OR NON-INFRINGEMENT.
 
 See the Apache 2 License for the specific language governing permissions and limitations under the License.
 */
-package com.msopentech.thali.universal.toronionproxy
+package com.msopentech.thali.toronionproxy
 
 import net.freehaven.tor.control.EventHandler
 import org.slf4j.Logger
