+++ android/src/main/java/com/msopentech/thali/toronionproxy/TorConfig.kt
@@ -10,7 +10,7 @@ MERCHANTABLITY OR NON-INFRINGEMENT.
 
 See the Apache 2 License for the specific language governing permissions and limitations under the License.
 */
-package com.msopentech.thali.universal.toronionproxy
+package com.msopentech.thali.toronionproxy
 
 import java.io.File
 import java.io.IOException
