+++ android/src/main/java/com/msopentech/thali/toronionproxy/TorControlConnection.kt
@@ -1,4 +1,4 @@
-package com.msopentech.thali.universal.toronionproxy
+package com.msopentech.thali.toronionproxy
 
 import net.freehaven.tor.control.TorControlConnection
 import java.io.*
