+++ android/src/main/java/com/msopentech/thali/toronionproxy/broadcaster/BaseEventBroadcaster.kt
@@ -10,8 +10,10 @@ MERCHANTABLITY OR NON-INFRINGEMENT.
 
 See the Apache 2 License for the specific language governing permissions and limitations under the License.
 */
-package com.msopentech.thali.universal.toronionproxy
+package com.msopentech.thali.toronionproxy.broadcaster
 
+import com.msopentech.thali.toronionproxy.settings.DefaultSettings
+import com.msopentech.thali.toronionproxy.settings.TorSettings
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import java.io.PrintWriter
