+++ android/src/main/java/com/msopentech/thali/toronionproxy/broadcaster/DefaultEventBroadcaster.kt
@@ -10,7 +10,9 @@ MERCHANTABLITY OR NON-INFRINGEMENT.
 
 See the Apache 2 License for the specific language governing permissions and limitations under the License.
 */
-package com.msopentech.thali.universal.toronionproxy
+package com.msopentech.thali.toronionproxy.broadcaster
+
+import com.msopentech.thali.toronionproxy.settings.TorSettings
 
 /**
  * Default EventBroadcaster. broadcastBandwidth, broadcastLogMessage, broadcastStatus are all
