+++ android/src/main/java/com/msopentech/thali/toronionproxy/settings/DefaultSettings.kt
@@ -1,4 +1,4 @@
-package com.msopentech.thali.universal.toronionproxy
+package com.msopentech.thali.toronionproxy.settings
 
 import java.util.*
