+++ android/src/main/java/com/msopentech/thali/toronionproxy/settings/TorSettingsBuilder.kt
@@ -10,8 +10,9 @@ MERCHANTABLITY OR NON-INFRINGEMENT.
 
 See the Apache 2 License for the specific language governing permissions and limitations under the License.
 */
-package com.msopentech.thali.universal.toronionproxy
+package com.msopentech.thali.toronionproxy.settings
 
+import com.msopentech.thali.toronionproxy.OnionProxyContext
 import java.io.*
 import java.net.InetSocketAddress
 import java.net.Socket
