+++ android/src/main/java/com/msopentech/thali/toronionproxy/util/Utilities.kt
@@ -10,7 +10,7 @@ MERCHANTABLITY OR NON-INFRINGEMENT.
 
 See the Apache 2 License for the specific language governing permissions and limitations under the License.
 */
-package com.msopentech.thali.universal.toronionproxy
+package com.msopentech.thali.toronionproxy.util
 
 import java.io.DataInputStream
 import java.io.DataOutputStream
